+++ src/domain/SummerHouse.java
@@ -1,7 +1,7 @@
 package domain;
 
 public class SummerHouse extends Residence{
-    public SummerHouse(double price, int squareMeter, int roomNumber, int saloonNumber) {
+    public SummerHouse(int price, int squareMeter, int roomNumber, int saloonNumber) {
         super(price, squareMeter, roomNumber, saloonNumber);
     }
 }
