+++ src/factory/ResidenceDTO.java
@@ -1,4 +1,4 @@
 package factory;
 
-record ResidenceDTO(double price, int squareMeter, int roomNumber, int saloonNumber) {};
+record ResidenceDTO(int price, int squareMeter, int roomNumber, int saloonNumber) {};
